Stop details RV flashing when updated
